requirements.txt
openai-whisper
flask
flask-cors